requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# PDF Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
pdf2image>=1.16.3
Pillow>=10.0.0

# OCR and Image Processing
pytesseract>=0.3.10
easyocr>=1.7.0

# ML/AI Models
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2

# Vector Store
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-groq>=0.0.1

# Utilities
numpy>=1.24.0
pandas>=2.0.0
